pyproject.toml
[project]
name = "acm-meeting-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "selenium>=4.40.0",
]
